+++ pmd-core/src/main/java/net/sourceforge/pmd/lang/AbstractParser.java
@@ -5,6 +5,10 @@
 package net.sourceforge.pmd.lang;
 
 import java.io.Reader;
+import java.nio.file.Paths;
+
+import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.ast.xpath.internal.FileNameXPathFunction;
 
 /**
  * This is a generic implementation of the Parser interface.
@@ -34,4 +38,13 @@ public abstract class AbstractParser implements Parser {
     }
 
     protected abstract TokenManager createTokenManager(Reader source);
+
+    @Deprecated
+    public static Node doParse(Parser parser, String fileName, Reader source) {
+        Node rootNode = parser.parse(fileName, source);
+        // remove prefixed path segments.
+        String simpleFileName = Paths.get(fileName).getFileName().toString();
+        rootNode.getUserMap().set(FileNameXPathFunction.FILE_NAME_KEY, simpleFileName);
+        return rootNode;
+    }
 }